initial commit from https://github.com/ixartz/Next-js-Blog-Boilerplate
this repo is based upon the boilerplate from
`ixartz`,`Next-js-Blog-Boilerplate`.

this commit is squashed down from that.

originally I followed readme and ran `git clone --depth=1`, but this
caused issues with pushing up onto a new repo, so I fixed it by deleting
the `.git` directory and re-initialising, with this commit as the first.

src: https://stackoverflow.com/a/29748753/827129