fix RTU working across page route changes by moving into `_app.tsx`
when this was mounted in `templates/Main.tsx` it was causing it to
remount each page load, but the script wouldn't re-initiate, so the
`_init` function in the script wasn't subsequently called, causing it to
fail to find the required elements to update.

this change persists the component across page route changes, which
keeps it working when the user navigates around the site.